requirements.txt
algoliasearch==3.0.0
algoliasearch-django==3.0.0
asgiref==3.8.1
boto3==1.35.30
botocore==1.35.30
certifi==2024.8.30
charset-normalizer==3.3.2
cloudinary==1.41.0
Django==5.1.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==30.3.0
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
razorpay==1.4.2
requests==2.32.3
s3transfer==0.10.2
setuptools==75.2.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
